seperated ui logic
