+++ src/components/ui/Navigation.css
@@ -1,0 +1,32 @@
+/* Simple Navigation */
+.simple-nav {
+  position: fixed;
+  top: 2rem;
+  right: 2rem;
+  z-index: 10;
+  display: flex;
+  gap: 2rem;
+}
+
+.nav-item {
+  background: none;
+  border: none;
+  color: white;
+  font-family: 'Instrument Sans', sans-serif;
+  font-size: 1.2rem;
+  font-weight: bold;
+  cursor: pointer;
+  padding: 0.5rem 0;
+  transition: opacity 0.3s ease;
+  opacity: 0.7;
+  text-decoration: underline;
+  text-underline-offset: 8px;
+}
+
+.nav-item:hover {
+  opacity: 1;
+}
+
+.nav-item.active {
+  opacity: 1;
+}
