+++ src/components/ui/SocialButtons.css
@@ -1,0 +1,44 @@
+/* Bottom Left Buttons */
+.bottom-left-buttons {
+  position: fixed;
+  bottom: 2rem;
+  left: 2rem;
+  z-index: 10;
+  display: flex;
+  align-items: center;
+  gap: 2rem;
+}
+
+.resume-button {
+  background: none;
+  border: none;
+  color: white;
+  font-family: 'Monda', sans-serif;
+  font-size: 2.2rem;
+  font-weight: bold;
+  cursor: pointer;
+  padding: 0.5rem 0;
+  transition: opacity 0.3s ease;
+  opacity: 0.9;
+}
+
+.resume-button:hover {
+  opacity: 1;
+}
+
+.icon-button {
+  background: none;
+  border: none;
+  color: white;
+  cursor: pointer;
+  padding: 0.5rem;
+  transition: opacity 0.3s ease;
+  opacity: 0.7;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.icon-button:hover {
+  opacity: 1;
+}
